move row to rowitem
