+++ shared/chat/inbox/index.stories.tsx
@@ -5,7 +5,13 @@ import * as Sb from '../../stories/storybook'
 import {isDarwin} from '../../constants/platform'
 import {isMobile, globalColors, globalMargins} from '../../styles'
 import Inbox from '.'
-import {RowItemSmall, RowItemBigHeader, RowItemBig, RowItemDivider, RowItemTeamBuilder} from './row/types.d'
+import {
+  RowItemSmall,
+  RowItemBigHeader,
+  RowItemBig,
+  RowItemDivider,
+  RowItemTeamBuilder,
+} from '../../constants/types/chat2/rowitem'
 import * as RPCChatTypes from '../../constants/types/rpc-chat-gen'
 
 /*
